게시판 적용 완료

+++ project04/src/main/java/kr/ed/haebeop/controller/BoardCtrl.java
@@ -4,11 +4,10 @@ import kr.ed.haebeop.domain.*;
 import kr.ed.haebeop.service.*;
 import kr.ed.haebeop.util.BoardPage;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.ServletContext;
@@ -73,6 +72,19 @@ public class BoardCtrl {
         model.addAttribute("page", page);
         model.addAttribute("curPage", curPage);
         List<BoardVO> boardList = boardService.boardList(page);
+
+        for(BoardVO boardVO : boardList) {
+            String authorNm = boardVO.getNm();
+            if(!authorNm.equals("관리자")) {
+                String nm = authorNm.substring(0, 1);
+                for(int i = 0; i < authorNm.length()-2; i++){
+                    nm += "*";
+                }
+                nm += authorNm.substring(authorNm.length() - 1);
+                boardVO.setNm(nm);
+            }
+        }
+
         model.addAttribute("boardList", boardList);
 
         BoardMgn boardMgn = boardMgnService.getBoardMgn(bmNo);
@@ -127,7 +139,7 @@ public class BoardCtrl {
                 String originalFilename = multipartFile.getOriginalFilename();
                 UUID uuid = UUID.randomUUID();
                 String uploadFilename = uuid.toString() + "_" + originalFilename;
-                
+
                 FileDTO fileDTO = new FileDTO();
                 fileDTO.setPar(bno);
                 fileDTO.setSaveFolder(String.valueOf(uploadPath));
@@ -155,11 +167,21 @@ public class BoardCtrl {
         int bno = Integer.parseInt(request.getParameter("bno"));
 
         BoardVO board = boardService.boardGet(bno, sid);
+        String bNm = board.getNm();
+
+        if(!bNm.equals("관리자")) {
+            String nm = bNm.substring(0, 1);
+            for(int i = 0; i < bNm.length()-2; i++){
+                nm += "*";
+            }
+            nm += bNm.substring(bNm.length() - 1);
+            board.setNm(nm);
+        }
         model.addAttribute("board", board);
 
         // 권한 관련 - 수정
         boolean addCheck = false;
-        if(sid != null && board.getAuthor().equals(sid)) {
+        if(!sid.equals("") && board.getAuthor().equals(sid)) {
             addCheck = true;
         }
         model.addAttribute("addCheck", addCheck);
@@ -265,35 +287,43 @@ public class BoardCtrl {
     }
 
     @GetMapping("/delete.do")
-    public String boardDeletePro(HttpServletRequest request) throws Exception {
+    public String boardDeletePro(HttpServletRequest request, RedirectAttributes rttr) throws Exception {
         String sid = session.getAttribute("sid") != null ? (String) session.getAttribute("sid") : "";
         int bno = Integer.parseInt(request.getParameter("bno"));
 
         BoardVO boardVO = boardService.boardGet(bno, sid);
-        int bmNo = boardVO.getBmNo();
 
-        FileDTO fileDTO = new FileDTO();
-        fileDTO.setPar(bno);
-        fileDTO.setToUse(toUseFileByBoard);
-        List<FileDTO> fileList = filesService.fileListByPar(fileDTO);
-        for(FileDTO files : fileList) {
-            File file = new File(files.getSaveFolder() + File.separator + files.getSaveNm());
-            if (file.exists()) {
-                file.delete();
-                filesService.filesDeleteAll(bno);
+        if(sid.equals(boardVO.getAuthor()) || sid.equals("admin")) {
+            int bmNo = boardVO.getBmNo();
+            FileDTO fileDTO = new FileDTO();
+            fileDTO.setPar(bno);
+            fileDTO.setToUse(toUseFileByBoard);
+            List<FileDTO> fileList = filesService.fileListByPar(fileDTO);
+            for(FileDTO files : fileList) {
+                File file = new File(files.getSaveFolder() + File.separator + files.getSaveNm());
+                if (file.exists()) {
+                    file.delete();
+                    filesService.filesDeleteAll(bno);
+                }
             }
-        }
 
-        commentService.commentDeleteAll(bno);
-        boardService.boardDelete(bno);
-        return "redirect:/board/list.do?no=" + bmNo;
+            commentService.commentDeleteAll(bno);
+            boardService.boardDelete(bno);
+            return "redirect:/board/list.do?no=" + bmNo;
+        } else {
+            rttr.addFlashAttribute("msg", "fail");
+            return "redirect:/board/get.do?bno=" + bno;
+        }
     }
 
     @PostMapping("commentAdd.do")
     @ResponseBody
-    public CommentVO commentInsert(@RequestBody Comment comment) throws Exception {
+    public CommentVO commentInsert(@RequestParam("par") int par, @RequestParam("content") String content) throws Exception {
         String sid = session.getAttribute("sid") != null ? (String) session.getAttribute("sid") : "";
+        Comment comment = new Comment();
         comment.setAuthor(sid);
+        comment.setPar(par);
+        comment.setContent(content);
         CommentVO commentVO = commentService.commentInsert(comment);
         String originNm = commentVO.getNm();
         if(!originNm.equals("관리자")) {
@@ -309,14 +339,25 @@ public class BoardCtrl {
 
     @PostMapping("commentRemove.do")
     @ResponseBody
-    public boolean commentDelete(@RequestBody Comment comment) throws Exception {
+    public boolean commentDelete(@RequestParam("cno") int cno) throws Exception {
         boolean result = false;
         String sid = session.getAttribute("sid") != null ? (String) session.getAttribute("sid") : "";
-        CommentVO commentVO = commentService.comment(comment.getCno());
+        CommentVO commentVO = commentService.comment(cno);
         if(commentVO.getAuthor().equals(sid) || sid.equals("admin")) {
             commentService.commentDelete(commentVO.getCno());
+            result = true;
         }
         return result;
+    }
+
+    @PostMapping("answerAdd.do")
+    @ResponseBody
+    public String answerInsert(@RequestParam("bno") int bno, @RequestParam("answer") String answer) throws Exception {
+        Board board = new Board();
+        board.setBno(bno);
+        board.setAnswer(answer);
+        boardService.qnaUpdate(board);
+        return answer;
     }
 
 }
